rename: 리뷰 관련 파일을 공통 폴더로 옮김
