+++ client/src/components/Review/Pagenation.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { FlexContainer } from 'pages/Review/ReviewPage';
+
+function Pagenation() {
+  return (
+    <FlexContainer>
+      <div>11232123</div>
+    </FlexContainer>
+  );
+}
+
+export default Pagenation;
