+++ client/src/components/Review/SearchBar.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable react/no-array-index-key */
 import GlobalStyle from 'GlobalStyles';
 import styled from 'styled-components';
-import { FlexContainer } from 'Review/Page/ReviewPage';
+import { FlexContainer } from 'pages/Review/ReviewPage';
 import { useState } from 'react';
 import SubjectMenu from './SubjectMenu';
